Moved my testing snippet into main.py
