--- test.py
@@ -1,17 +1,0 @@
-import ast_app
-from ast_app.core.data_adapters.kml.adapter import KMLAdapter
-
-adapter = KMLAdapter()
-gdf = adapter.read(path="ast_app/Test_Shape_A/Test_Shape_A.kmz", target_crs="EPSG:3005")
-
-print(type(gdf))
-print(gdf.shape)
-print(gdf.crs)
-print(gdf.head())
-
-def test_kml_adapter_reads_file():
-    adapter = KMLAdapter()
-    gdf = adapter.read(path="ast_app/Test_Shape_A/Test_Shape_A.kmz")
-    assert gdf is not None
-    assert gdf.crs is not None
-    assert len(gdf) > 0
